+++ packages/faas-performance/src/classes/requestPermLog.js
@@ -1,13 +1,14 @@
-const nanoid = require('nanoid');
 const moment = require('moment');
-const { getRequestPermCollectionName } = require('../utils');
+const { getRequestPermCollectionName } = require('../utils/collection');
 
 class RequestPermLog {
-  constructor(db, lambdaId) {
-    this.db = db.collection(getRequestPermCollectionName(lambdaId));
+  constructor(app, db, lambdaId) {
+    this.collection = db.collection(getRequestPermCollectionName(lambdaId));
+    if (app.tigo.faas.perm.maxKeepDays) {
+      this.collection.ensureIndex({ point: 1 }, { expireAfterSeconds: maxKeepDays * 86400 });
+    }
   }
   begin() {
-    this.executionId = nanoid();
     this.startTime = Date.now();
   }
   async end() {
@@ -15,13 +16,13 @@ class RequestPermLog {
     this.executionTime = this.endTime - this.startTime;
     // calc avg per min
     const point = moment().startOf('minute').valueOf();
-    const stored = await this.db.findOne({
+    const stored = await this.collection.findOne({
       point,
     });
-    if (!stored) {
-      stored.avgTimeCost = (avgTimeCost * stored.count + this.executionTime) / (stored.count + 1);
+    if (stored) {
+      stored.avgTimeCost = (stored.avgTimeCost * stored.count + this.executionTime) / (stored.count + 1);
       stored.count += 1;
-      await this.db.updateOne(
+      await this.collection.updateOne(
         {
           point,
         },
@@ -30,7 +31,7 @@ class RequestPermLog {
         }
       );
     } else {
-      await this.db.insertOne({
+      await this.collection.insertOne({
         point,
         avgTimeCost: this.executionTime,
         count: 1,