+++ packages/faas-performance/src/utils/collection.js
@@ -1,5 +1,7 @@
 const getRequestPermCollectionName = (lambdaId) => `${lambdaId}_request`;
+const getRequestStatusCollectionName = (lambdaId) => `${lambdaId}_request_status`;
 
 module.exports = {
   getRequestPermCollectionName,
+  getRequestStatusCollectionName,
 };