+++ packages/faas-log/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tigojs/faas-log changelog
 
+## v0.2.0
+
+- Breaking change: `/faas/queryLogs` now changed to `/faas/log/list`.
+
+- Feat: add `maxKeepDays` option.
+
+- Minor: upgrade `@tigojs/utils`.
+
 ## v0.1.2 - v0.1.3
 
 - Minor: upgrade `@tigojs/utils`.