+++ packages/mysql/CHANGELOG.md
@@ -1,5 +1,5 @@
 # @tigojs/mysql changelog
 
-## v0.1.28
+## v0.1.28 - v0.1.29
 
 - Minor: upgrade `@tigojs/utils`