+++ packages/oss/CHANGELOG.md
@@ -1,5 +1,5 @@
 # @tigojs/oss changelog
 
-## v0.1.27
+## v0.1.27 - v0.1.28
 
 - Minor: upgrade `@tigojs/utils`.