+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @tigojs/utils changelog
 
+## v0.5.4
+
+- Fix: `registerController` method will detect if target existed now.
+
 ## v0.5.2
 
 - Feat: add `parseContextQuery` method to parse numbers in the `ctx.query`.